Moved history API calls into a service like all the others
